Normalize matched/ignored paths
Ignore patterns and path strings are normalized for comparison:

- path separators replaced with slashes
- characters are lower cased

The original logic of recording comparison matches
(unmodified) in the pathsMatched runtime collection
for later use is retained.

NOTE:

This commit also rearranges files in the
`internal/restart/files` package to match the
`internal/restart/registry` package.
