+++ internal/restart/files/assertions_windows.go
@@ -1,5 +1,5 @@
-//go:build !windows
-// +build !windows
+//go:build windows
+// +build windows
 
 // Copyright 2022 Adam Chalkley
 //
@@ -8,31 +8,45 @@
 // Licensed under the MIT License. See LICENSE file in the project root for
 // full license information.
 
-package registry
-
-// NOTE: This package is not intended for non-Windows systems.
+package files
 
 import (
 	"github.com/atc0005/check-restart/internal/restart"
 )
 
 // DefaultRebootRequiredIgnoredPaths provides the default collection of paths
-// for registry related reboot required assertions that should be ignored.
+// for file related reboot required assertions that should be ignored.
 //
 // Paths are normalized before comparison with matched paths.
 //
 // For consistency, these entries should match the default path syntax for the
 // operating system in question.
 func DefaultRebootRequiredIgnoredPaths() []string {
-
-	logger.Println("WARNING: This tool is not supported for non-Windows systems!")
 	return []string{}
 }
 
-// DefaultRebootRequiredAssertions provides the default collection of registry
+// DefaultRebootRequiredAssertions provides the default collection of file
 // related reboot required assertions.
 func DefaultRebootRequiredAssertions() restart.RebootRequiredAsserters {
 
-	logger.Println("WARNING: This tool is not supported for non-Windows systems!")
-	return restart.RebootRequiredAsserters{}
+	var assertions = restart.RebootRequiredAsserters{
+
+		// Test entry.
+		// &File{
+		// 	// path: `C:\Windows\notepad.exe`,
+		// 	envVarPathPrefix: "SystemRoot",
+		// 	path:             `notepad.exe`,
+		// },
+
+		// Found on Windows desktop and server variants after applying Windows
+		// Updates.
+		&File{
+			// path: `C:\Windows\WinSxS\pending.xml`,
+			envVarPathPrefix: "SystemRoot",
+			path:             `WinSxS\pending.xml`,
+		},
+	}
+
+	return assertions
+
 }
